requirements.txt
asgiref==3.8.1
azure-core==1.32.0
azure-storage-blob==12.24.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
coverage==7.6.9
cryptography==44.0.0
Django==5.1.2
django-cors-headers==4.6.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
isodate==0.7.2
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
pytest-mock==3.14.0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
tzdata==2024.2
whitenoise==6.8.2
